move survey export top-level declarations
