+++ posthog-node/package.json
@@ -1,6 +1,6 @@
 {
   "name": "posthog-node",
-  "version": "4.11.5",
+  "version": "4.11.6",
   "description": "PostHog Node.js integration",
   "repository": {
     "type": "git",
