+++ posthog-react-native/src/surveys/components/SurveyModal.tsx
@@ -15,7 +15,7 @@ import { ConfirmationMessage } from './ConfirmationMessage'
 import { Questions } from './Surveys'
 
 import { SurveyAppearanceTheme } from '../surveys-utils'
-import { Survey, SurveyQuestionDescriptionContentType } from '../../../../posthog-core/src/surveys-types'
+import { Survey, SurveyQuestionDescriptionContentType } from '../../../../posthog-core/src'
 import { useOptionalSafeAreaInsets } from '../../optional/OptionalReactNativeSafeArea'
 
 export type SurveyModalProps = {
