+++ posthog-react-native/src/surveys/components/Surveys.tsx
@@ -2,7 +2,7 @@ import React, { useMemo, useState } from 'react'
 import { ScrollView, StyleProp, ViewStyle } from 'react-native'
 
 import { getDisplayOrderQuestions, SurveyAppearanceTheme } from '../surveys-utils'
-import { Survey, SurveyAppearance, SurveyQuestion } from '../../../../posthog-core/src/surveys-types'
+import { Survey, SurveyAppearance, SurveyQuestion } from '../../../../posthog-core/src'
 import { LinkQuestion, MultipleChoiceQuestion, OpenTextQuestion, RatingQuestion } from './QuestionTypes'
 import { PostHog } from '../../posthog-rn'
 import { usePostHog } from '../../hooks/usePostHog'
