+++ posthog-react-native/src/surveys/index.ts
@@ -2,5 +2,5 @@ export { PostHogSurveyProvider } from './PostHogSurveyProvider'
 export type { PostHogSurveyProviderProps } from './PostHogSurveyProvider'
 export { SurveyModal } from './components/SurveyModal'
 export type { SurveyModalProps } from './components/SurveyModal'
-export type { Survey, SurveyAppearance } from '../../../posthog-core/src/surveys-types'
+export type { Survey, SurveyAppearance } from '../../../posthog-core/src'
 export { Questions } from './components/Surveys'
